refactor(web): tighten frontend state budget debt
