+++ web/src/lib/features/chat/project-conversation-controller.svelte.ts
@@ -4,10 +4,8 @@ import type { ProjectAIFocus } from './project-ai-focus'
 import { createProjectConversationControllerApi } from './project-conversation-controller-api'
 import { projectConversationHasPendingInterrupt } from './project-conversation-controller-helpers'
 import { createProjectConversationControllerOperations } from './project-conversation-controller-operations'
-import {
-  createProjectConversationControllerActions,
-  sendNextQueuedProjectConversationTurn,
-} from './project-conversation-controller-actions'
+import { createProjectConversationControllerActions } from './project-conversation-controller-actions'
+import { createProjectConversationQueuedTurnDispatcher } from './project-conversation-controller-dispatch'
 import {
   canQueueProjectConversationTurn,
   createProjectConversationTabState,
@@ -37,8 +35,6 @@ export function createProjectConversationController(
   let nextTabID = 0
   let nextQueuedTurnID = 0
   let snapshotNotificationQueued = false
-  let queuedTurnDispatchScheduled = false
-  const autoDispatchQueuedTurnIDByTab = new Map<string, string>()
 
   function touch() {
     revision += 1
@@ -162,63 +158,14 @@ export function createProjectConversationController(
     touch,
     operations,
   })
+  const queuedTurnDispatcher = createProjectConversationQueuedTurnDispatcher({
+    getTabs: () => tabs,
+    sendTurnInTab: operations.sendTurnInTab,
+    touch,
+  })
 
   function scheduleQueuedTurnDispatch() {
-    if (queuedTurnDispatchScheduled) {
-      return
-    }
-
-    queuedTurnDispatchScheduled = true
-    queueMicrotask(() => {
-      queuedTurnDispatchScheduled = false
-
-      for (const tab of tabs) {
-        const nextQueuedTurnId = tab.queuedTurns[0]?.id ?? ''
-        const shouldAutoDispatch =
-          !!nextQueuedTurnId &&
-          !!tab.projectId &&
-          !!tab.providerId &&
-          tab.phase === 'idle' &&
-          !projectConversationHasPendingInterrupt(tab.entries)
-
-        if (!shouldAutoDispatch) {
-          autoDispatchQueuedTurnIDByTab.delete(tab.id)
-          continue
-        }
-
-        if (autoDispatchQueuedTurnIDByTab.get(tab.id) === nextQueuedTurnId) {
-          continue
-        }
-
-        autoDispatchQueuedTurnIDByTab.set(tab.id, nextQueuedTurnId)
-        queueMicrotask(() => {
-          const liveTab = tabs.find((item) => item.id === tab.id) ?? null
-          if (
-            !liveTab ||
-            liveTab.phase !== 'idle' ||
-            projectConversationHasPendingInterrupt(liveTab.entries) ||
-            (liveTab.queuedTurns[0]?.id ?? '') !== nextQueuedTurnId
-          ) {
-            if (autoDispatchQueuedTurnIDByTab.get(tab.id) === nextQueuedTurnId) {
-              autoDispatchQueuedTurnIDByTab.delete(tab.id)
-            }
-            return
-          }
-
-          void sendNextQueuedProjectConversationTurn({
-            tab: liveTab,
-            sendTurnInTab: operations.sendTurnInTab,
-          }).then((sent) => {
-            if (autoDispatchQueuedTurnIDByTab.get(tab.id) === nextQueuedTurnId) {
-              autoDispatchQueuedTurnIDByTab.delete(tab.id)
-            }
-            if (sent) {
-              touch()
-            }
-          })
-        })
-      }
-    })
+    queuedTurnDispatcher.schedule()
   }
 
   ensureTabExists()
