+++ web/src/lib/features/chat/terminal-manager.svelte.ts
@@ -1,16 +1,10 @@
-import {
-  encodeTerminalPayload,
-  mountProjectConversationTerminal,
-} from './project-conversation-terminal-panel-helpers'
 import { createTerminalConnectionHelpers } from './terminal-manager-connection'
 import {
-  TERMINAL_RECONNECT_ATTEMPT_LIMIT,
   clearTerminalReconnectTimer,
   ensureTerminalRuntime,
-  forgetTerminalRuntime,
   generateTerminalManagerID,
-  nextTerminalReconnectDelay,
 } from './terminal-manager-runtime'
+import { createTerminalLifecycleHelpers } from './terminal-manager-lifecycle'
 import type {
   MountedTerminal,
   TerminalInstance,
@@ -44,6 +38,18 @@ export function createTerminalManager(input: {
     return instances.some((inst) => inst.id === id)
   }
 
+  const { closeSocket, mountTerminal, scheduleReconnect, unmountTerminal } =
+    createTerminalLifecycleHelpers({
+      runtimeMap,
+      socketMap,
+      xtermMap,
+      elementMap,
+      resizeObserverMap,
+      hasInstance,
+      updateInstance,
+      connectTerminal,
+    })
+
   const { attachSocket, matchesConnectionState, resolveTerminalSession, setConnectingStatus } =
     createTerminalConnectionHelpers({
       getConversationId: input.getConversationId,
@@ -54,138 +60,6 @@ export function createTerminalManager(input: {
       scheduleReconnect,
       updateInstance,
     })
-
-  async function mountTerminal(id: string, element: HTMLDivElement) {
-    // Prevent double-mount
-    if (xtermMap.has(id) && elementMap.get(id) === element) return
-
-    const runtime = ensureTerminalRuntime(runtimeMap, id)
-    runtime.mountRevision += 1
-    const mountRevision = runtime.mountRevision
-
-    // Clean up previous mount for this id
-    unmountTerminal(id, false)
-    elementMap.set(id, element)
-
-    const mounted = await mountProjectConversationTerminal({
-      element,
-      onData: (data) => {
-        const socket = socketMap.get(id)
-        if (socket?.readyState !== WebSocket.OPEN) return
-        socket.send(JSON.stringify({ type: 'input', data: encodeTerminalPayload(data) }))
-      },
-      onResize: ({ cols, rows }) => {
-        const socket = socketMap.get(id)
-        if (socket?.readyState !== WebSocket.OPEN) return
-        socket.send(JSON.stringify({ type: 'resize', cols, rows }))
-      },
-    })
-
-    if (
-      !hasInstance(id) ||
-      runtimeMap.get(id)?.mountRevision !== mountRevision ||
-      elementMap.get(id) !== element
-    ) {
-      mounted.dispose()
-      return
-    }
-
-    xtermMap.set(id, mounted)
-
-    const ro = new ResizeObserver(() => {
-      const entry = xtermMap.get(id)
-      if (!entry) return
-      entry.fitAddon.fit()
-      const socket = socketMap.get(id)
-      if (socket?.readyState === WebSocket.OPEN) {
-        socket.send(
-          JSON.stringify({ type: 'resize', cols: entry.terminal.cols, rows: entry.terminal.rows }),
-        )
-      }
-    })
-    ro.observe(element)
-    resizeObserverMap.set(id, ro)
-  }
-
-  function unmountTerminal(id: string, forget: boolean) {
-    resizeObserverMap.get(id)?.disconnect()
-    resizeObserverMap.delete(id)
-    closeSocket(id, { updateStatus: false, reconnect: false, terminate: true })
-    xtermMap.get(id)?.dispose()
-    xtermMap.delete(id)
-    elementMap.delete(id)
-    if (forget) {
-      forgetTerminalRuntime(runtimeMap, id)
-    }
-  }
-
-  function closeSocket(
-    id: string,
-    options: {
-      updateStatus: boolean
-      reconnect: boolean
-      terminate: boolean
-    },
-  ) {
-    const runtime = ensureTerminalRuntime(runtimeMap, id)
-    runtime.reconnectEnabled = options.reconnect
-    clearTerminalReconnectTimer(runtimeMap, id)
-
-    const socket = socketMap.get(id)
-    socketMap.delete(id)
-    if (socket?.readyState === WebSocket.OPEN) {
-      if (options.terminate) {
-        socket.send(JSON.stringify({ type: 'close' }))
-      }
-      socket.close()
-    } else if (socket?.readyState === WebSocket.CONNECTING) {
-      socket.close()
-    }
-    if (options.terminate) {
-      runtime.session = null
-    }
-    if (options.updateStatus) {
-      updateInstance(id, { status: 'closed', statusMessage: 'Terminal closed.', sessionID: '' })
-    }
-  }
-
-  function scheduleReconnect(id: string, label: string) {
-    const runtime = runtimeMap.get(id)
-    if (
-      !runtime ||
-      !runtime.reconnectEnabled ||
-      !runtime.session ||
-      !hasInstance(id) ||
-      !xtermMap.has(id)
-    ) {
-      return
-    }
-
-    if (runtime.reconnectAttempts >= TERMINAL_RECONNECT_ATTEMPT_LIMIT) {
-      runtime.reconnectEnabled = false
-      updateInstance(id, {
-        status: 'error',
-        statusMessage: 'Terminal disconnected. Reconnect attempts exhausted.',
-        sessionID: '',
-      })
-      return
-    }
-
-    runtime.reconnectAttempts += 1
-    const delay = nextTerminalReconnectDelay(runtime.reconnectAttempts)
-    updateInstance(id, {
-      status: 'connecting',
-      statusMessage: `Reconnecting shell in ${label}...`,
-      sessionID: '',
-    })
-    runtime.reconnectTimer = setTimeout(() => {
-      runtime.reconnectTimer = null
-      if (!runtime.reconnectEnabled || !hasInstance(id) || !xtermMap.has(id)) {
-        return
-      }
-      void connectTerminal(id, true)
-    }, delay)
-  }
 
   async function connectTerminal(id: string, isReconnect = false) {
     const conversationId = input.getConversationId()
